+++ modules/util/util.js
@@ -195,8 +195,9 @@ export function utilGetAllNodes(ids, graph) {
  *                            If falsy, it's for a label elsewhere in the UI.
  */
 export function utilDisplayName(entity, flags) {
-    var localizedNameKey = 'name:' + localizer.languageCode().toLowerCase();
-    var name = entity.tags[localizedNameKey] || entity.tags.name || '';
+    const name = localizer.expandedLocaleCodes()
+        .map(code => entity.tags[`name:${code}`])
+        .find(Boolean) || entity.tags.name || '';
 
     var tags = {
         direction: entity.tags.direction,